Added limit.h to the list of headers so that Guitar builds start working.
A macro called NAME_MAX is being used in the code, which is supposed to
originate in limits.h, which was previously not included in messaging.cc.
However, local builds still work, so I assume that it was being included
trasitively and it was just never an issue. However Guitar fails because it
can't find NAME_MAX, so there must be a subtle difference in the toolchain that
Guitar is using that results in this failure.

PiperOrigin-RevId: 244924017