Move UserSecureManager, UserSecureInternal, platform-specific storage, and associated tests from Auth Desktop to App.
Logic still needs to be added to allow systems to create a UserSecureManager
for a given domain (auth, instanceid) but for now this maintains the current
functionality of Auth persistence.

This also means that the libsecret customer has moved from Firebase Auth to
Firebase App, though usage of the library (via UserSecureManager) will not change.

Also updates open-source build.

PiperOrigin-RevId: 248002024