Desktop UUID generator class.
Windows adds the requirement that apps link Rpcrt4.lib which isn't
present in third_party/msvc at the moment so the test won't link.

Added disabled code for Linux that can be enabled to use libuuid if/when
we add it to third_party.  For the moment Linux uses /dev/urandom to generate
a random 128-bit number.

iOS / Darwin uses NSUUID which appears to work fine.

Android is not required at the moment and therefore not supported.

PiperOrigin-RevId: 248206226