Implement Firebase Instance ID on desktop to call into App's IID library.
Implement GetId() and GetToken() synchronously (for now). DeleteId() and DeleteToken() are also implemented synchronously (though they don't call Delete on the server yet).

Turns out that the AppId needs to use an alternate version of Base64 that is "url safe" so this CL adds support for that as well, and fleshes out the tests/fuzzer to ensure coverage.

PiperOrigin-RevId: 248639274