Add CMake logic for defining a future library for playbilling
Adds a cmake script that allows the definition of a version of the future library that uses the namespace playbillingclient instead of firebase. Adds logic to the source files that if playbilling is being used, it includes the generated public header for it instead.
Because of the source libraries needing to include the generated header, it will currently only work for playbilling, but could potentially be generalized in the future.

PiperOrigin-RevId: 251918893