Make firebase::kDefaultAppName const.
The Google C++ Style guide says that names starting with a leading "k"
should be for constants, i.e. for variables declared "constexpr" or "const"
and whose value is fixed for the duration of the program.  But before
this change, kDefaultAppName was not declared "const", and could be modified.

This change is not strictly backwards compatible; any programs which did attempt
to modify kDefaultAppName would break.  However, kDefaultAppName is not documented
in the API documentation, so I think this is OK.  Also a Google search for
"kDefaultAppName" and "firebase" didn't get any hits other than the Firebase API docs
and the Firebase source code, so don't think there is any open-source code which
is doing this.

PiperOrigin-RevId: 256182269