Open Source support for IID Tests, testing/cppsdk
This CL makes the changes necessary to run the basic Instance ID tests with the Firebase ++ Open Source build.  In order to make this work, this CL also adds Open Source support to //firebase/testing/cppsdk/, which is used by the IID tests (and presumably a number of other C++ tests).

Additionally, added a workflow to the copy.bara.sky script to support copying from in-progress CLs.

PiperOrigin-RevId: 258222273