Enable Storage OSS Tests (Desktop)
Moved test files to tests/ folder and added CMakeLists.txt file for tests.

Also added json_util.h/cc to testing/cppsdk to replace the EqualsJson functionality that was being used from the internal version of jsoncpp (in third_party) to avoid having that as a dependency for open source.

PiperOrigin-RevId: 258970591