Added ability to create a firebase::App for existing FIRApp/FirebaseApp.
iOS FIRApp and Android FirebaseApp objects are serialized and are
therefore restored when the application restarts. A firebase::App could
only be created if a FIRApp/FirebaseApp was not already present unless
the FirebaseApp was a default instance.  This prevented developers from
creating named firebase::App instances since upon application restart
it would not be possible to create a firebase::App as a FirebaseApp
would already exist with the same name. Furthermore, once a named
firebase::App object was destroyed it wasn't possible to recreate the
same object in the same session as a FIRApp/FirebaseApp with the same
name would be present.

This changes the behavior of firebase::App on iOS and Android to allow
connection with an existing FIRApp/FirebaseApp object. If the
FIRApp/FirebaseApp has different options than those requested by the
caller of firebase::App::Create(), the existing FIRApp/FirebaseApp object
is destroyed and recreated with the requested options.

PiperOrigin-RevId: 261147650