Fix tests incorrectly instancing firebase::App on desktop.
api_key and app_id are required options when creating a
firebase::App instance, this moves all tests to use a
header-only library that provides a mock AppOptions with
the required parameters set.  In addition, the header
library handles platform abstraction for the creation of
firebase::App in the test environment.
PiperOrigin-RevId: 261342319