Scripts to run the tests on desktop platforms with Kokoro
Adds the scripts to run CMake tests for all of the desktop platforms. Adds the logic needed to the initial Mac script to set up the environment as needed to run the tests. Fixes the CMake definitions of tests to build on Linux and Mac. Adds explicit casting of longs to int64_t in variant tests, as apparently on Mac that is required.

Note that Windows will still fail during the CMake step, as the dependencies are not yet set up.

PiperOrigin-RevId: 262447157