Auth: C++ API support for existing Generic IDP implementation on iOS.
C++ API implementation calls into the Objective C Firebase Auth iOS SDK.  Includes updates to the ios_testapp, integration tests, and a new ios_util function to handle converting Vectors of strings to NSMutableArrays.

Moved the GetProviderId functions of the C++ providers to the header so that the implementation is common across all targets.

Currently iOS does not support User linkWithProvider and reauthenticateWithProvider, but it's in the works.  For now the implementation gets a credential and the uses that to linke/reauthenticate before returning the SignInResult.

PiperOrigin-RevId: 262596855