Work on script to run CMake tests on Windows Kokoro
Add logic to the backend script to install absl python on Windows. Change Auth desktop to use the flatbuffers unique pointer for the flatbuffer files, which is what the other Firebase libraries do. Fix storage to use a flatbuffer unique pointer where one is needed. Run ctest in verbose mode, since by default it only outputs that a test failed, and not how it failed.

Note that the tests intentionally disable Database and Remote Config, as those require additional requirements (Openssl and Protobuf) that still need to be figured out.

PiperOrigin-RevId: 263453335