Auth: C++ API support for existing Generic IDP implementation on Android.
C++ API implementation calls into the Java Firebase Android iOS SDK.  Includes updates to the integration tests, and a new android_util function to handle converting Vectors of std::strings to ArrayLists.

Additionally added a switch to turn tests that require manual interactions to be disabled by default in the integration test suite so that it may run properly on CI systems.

Moved the GetProviderId functions of the C++ providers to the header so that the implementation is common across all targets.

Currently iOS does not support User linkWithProvider and reauthenticateWithProvider, but it's in the works.  For now the implementation gets a credential and the uses that to linke/reauthenticate before returning the SignInResult.

PiperOrigin-RevId: 263859771