Make FutureManager not delete futures if they are still referenced externally.
A recent change to reference counted future impl made IsSafeToDelete only check if the underlying logic determined if it was safe, which was needed when deleting Auth futures directly. This made FutureManager a bit hasty in deleting things, which instead should be checking if those Futures are still being referenced by someone else.

PiperOrigin-RevId: 264503406