Fixed some potential race conditions in DatabaseReference's various constructors.
Because the database keeps its own bookkeeping regarding the memory addresses
of the QueryInternals and DatabaseReferenceInternals, if they are created and added to
the Database's registry in two steps, a situation could arise where the
database shuts down with either stale or incomplete internal bookkeeping. To
solve this, any place that we update Database's bookkeeping we need do
so atomically.

PiperOrigin-RevId: 265136296