Automated g4 rollback of changelist 266264551.
*** Reason for rollback ***

This breaks Performance Monitoring and Remote Config iOS builds due to their dependency upon A/B testing.  The A/B testing framework no longer has a binary framework which means that the snapshot no longer works since it doesn't pull in a snapshot of the source :/

So _someone_ is going to have to unravel this so we can update pods for the next release.  Also it won't be able to build Firebase Installations from a snapshot of the iOS release (i.e will require the blaze flag "--define firebase_build=head") until this is resolved.

*** Original change description ***

Update pods to 6.7.0

***

PiperOrigin-RevId: 266272837