Automated g4 rollback of changelist 267236680.
*** Reason for rollback ***

Roll-forward (i.e. rollback of rollback) after fixing the breakage that caused the initial rollback.  This CL has three fixes:
(1) It incorporates the fixes by varconst@ in CL 267229592, which fixed a problem where
    some of the references to std::function were not inside
    "#if defined(FIREBASE_USE_STD_FUNCTION)".
(2) It deletes an unnecessary (and wrong) friend declaration that was causing an ambiguity error; there's no need to declare the nested class intrusive_list::intrusive_list_iterator as a friend, since the containing class intrusive_list is already a friend, and the nested class has the same access as the containing class. The invalid friend declaration was implicitly declaring a non-nested class of the same name, causing later references to that name to be ambiguous.
(3) It adds a missing "const" to the declarations of some copy-constructor and assignment operators (that are only declared, not defined).

This CL is best reviewed by ticking the "Diff against rollback base revisions" box in Critique.

*** Original change description ***

Automated g4 rollback of changelist 267110370.

*** Reason for rollback ***

Fails to compile on Windows.

Can test a fix via:
blaze build --config=msvc //firebase/app/client/cpp:app_kokoro

*** Original change description ***

Add support for multiple completion handlers to Future.

This includes support for both adding and removing callbacks,
including removing callbacks added as std::function<...>.
OnCompletion() does not remove completions added with AddCompletion().

The implementation uses...

***

PiperOrigin-RevId: 267398802