Automated g4 rollback of changelist 267625598.
*** Reason for rollback ***

Dependent on Rolling back CL 267398802 which breaks linux build

*** Original change description ***

Add blocking methods Wait() and Await() to FutureBase and Future
(respectively).  Wait() just blocks until the future is completed,
or a timeout expires.  Await() blocks until the future is completed,
or a timeout expires, and then returns a pointer to the future's result,
or nullptr if the timeout expired.

***

PiperOrigin-RevId: 267672880