Automated g4 rollback of changelist 267672880.
*** Reason for rollback ***

Rolling original CL forward again, because the CL that it depends on has been rolled forward again in CL 268529975 five days ago.

*** Original change description ***

Automated g4 rollback of changelist 267625598.

*** Reason for rollback ***

Dependent on Rolling back CL 267398802 which breaks linux build

*** Original change description ***

Add blocking methods Wait() and Await() to FutureBase and Future
(respectively).  Wait() just blocks until the future is completed,
or a timeout expires.  Await() blocks until the future is completed,
or a timeout expires, and then returns a pointer to the future's result,
or nullptr if the timeout expired.

***

***

PiperOrigin-RevId: 269523992