Handle the C++ SDK being built for Android without doing Gradle first
Add logic that allows CMake to invoke gradle if the dexed jar files are not present already. Defines the proguard files with the same logic, so that the Unity SDK, which needs to access the proguard file, can access it.

Currently the build logic of CMake invoking Gradle causes some parts of the build to be unnecessarily duplicated, and will need to be improved in the future.

PiperOrigin-RevId: 277102239