Automated g4 rollback of changelist 275555534.
*** Reason for rollback ***

The instance id desktop impl internally is hidden behind the EAP flag, and the correct usage of the library is to link against it, and use that flag to disable/enable it.  As such, hiding it behind another flag in CMake causes problems, as the potential users of the iid desktop impl always try using it, and assume it will handle being enabled internally.

*** Original change description ***

Hide the instance id desktop implementation behind a feature flag in open source

***

PiperOrigin-RevId: 280665206