Fixed exception when printing app IDs with python3.
Passing a byte array (output of str.encode()) to sys.stdout.write() in
python3 throws an exceeption so this changes
generate_xml_from_google_services_json.py to write an unencoded string
as sys.stdout.write() encodes the string for the output terminal
anyway.

In addition, this migrates generate_xml_from_google_services_json.py
to be compatible with python3, expands test cases for the script and adds
python2 and python3 testing to the presubmit.

PiperOrigin-RevId: 281637905