Initial implementation of the FTL game SDK for Android (C++)
Implementation of the API for Firebase Test Lab Game Loops in C++. Gathers
game loop data from the Android SDK via JNI calls.

Currently, the app is exited by killing the thread with a call to exit(). This
worked locally, but isn't the proper way to exit a native app. Simply
finishing the activity puts the app in a bad state where it can't be started
again. Also, custom results are currently unformatted.

WANT_LGTM=smiles

#FTLGameSDK

PiperOrigin-RevId: 291940340