Added NoopPersistenceManager, removed NoopPersistenceStorageEngine.
At some point I added a NoopPersistenceStorageEngine which is a version of the PersistenceStorageEngine interface which takes no actions, intended for use with the special .info sync tree which does not persist it's data. However, that was the wrong level of abstraction, and it was the PersistenceManager that should have had the Noop version of it. This fixes a potential issue where .info values are returned null instead of the value they are supposed to be.

PiperOrigin-RevId: 293012252