Make `FirestoreInternal` store a pointer to the corresponding public `Firestore` instance.
The existing setup instead stores a pointer to `App` and uses it to retrieve the corresponding `Firestore` from the "main" cache of Firestore instances. This is unnecessarily tricky and falls apart in tests, which use their own cache of Firestore instances not synchronized with the "main" cache. What happens is that a call to e.g. `DocumentReference::firestore()` tries to retrieve an instance by App from the main cache, but because the main cache is empty, it leads to creation of a new Firestore instance.

PiperOrigin-RevId: 295244529