Add performance to the release build files.
Performance is missing from the release build files, which prevents the EAP builds from testing with the testapp builder. Note that this applies to the non-EAP build as well, however will be non-functional, as the C++ library to link against is missing. Firestore has been in this state for a while, with no complaints.

PiperOrigin-RevId: 295859757