Use a single function, `FieldValue::Increment`, for both integer and double arguments.
See go/firestore-cpp-field-value-increment for more background and rationale.

This requires some template metaprogramming but results in a more convenient and safer API. In particular, `static_assert` allows preventing truncation with a readable error message (e.g. if a user passes `unsigned long long` as an argument).

While working on this, I discovered that STLPort actually does contain `<type_traits>` header, so the STLPort implementation is pretty straightforward (the only extra complication is accounting for the nested `tr1` namespace). I checked the code from this CL manually in Android Studio using STLPort on NDK r17, and it compiled successfully.

PiperOrigin-RevId: 308336176