Add content provider support to the FTL Android initialization
The FTL game loop runner was recently changed to add a content provider to share game loop data with the app under test (cl/307423902). This is a more robust way of passing game loop scenario data since the content provider can always be queried, but the intent that starts the game loop can be lost (e.g. when Unity transitions to a different scene). This adds support for the SDK to retrieve the data from the content provider at initialization, and fall back on the intent for backwards compatability.

#FTLGameSDK

PiperOrigin-RevId: 311586597