Change the signature of `RunTransaction` callback to `Error (Transaction&, std::string&)`.
Mutable references are preferable in this case to pointers because they clearly indicate that the given argument is always valid (not null), potentially saving our users the trouble of checking explicitly.

The C++ style guide now (cr/312129146) allows using mutable references as function parameters. Note that our users never invoke the callback (this is done by Firestore), so the absence of an ampersand at the call site will not affect readability for them in any way.

PiperOrigin-RevId: 312197508