Add `FirebaseFirestore.ListenForSnapshotsInSync()`
This is the first in a series that implements a new pattern for managing listener callbacks. Previously listeners needed to register a callback with C++ and maintain a dictionary of callbacks keyed off of a unique id and cleaned up once listening stopped or the firestore instance was disposed. The pattern wasn't implemented properly and had to be redone.

This CL reimplements the pattern by:
1. Introducing a helper class, `ListenerRegistrationMap`, to keep a dictionary of all the callbacks. A static instance of this will get introduced for each `Listen*` method.
1. Updating `ListenerRegistration` to unregister callbacks when `Stop()` is called.
1. The C++ listener class that accepts the callback from C# will receive it in the native listen method instead of requiring a global callback registration.
1. In a future CL all the callbacks attached to a specific firestore instance will be pruned from its `ListenerRegistrationMap` instance.

Once this pattern has been approved in this CL it will get applied to other `Listen*` methods in the API.

PiperOrigin-RevId: 312713181