Fix most Unity leaks caused by incorrect SWIG configuration.
There were two major sources of leaks:
- SWIG was configured to ignore destructors of the public C++ classes; consequently, all of them were leaked;
- SWIG was leaking the standard containers (and all their elements).

The former problem is solved by fixing the SWIG configuration (as outlined in https://docs.google.com/document/d/1qqYXGzbB-01l00Mv-dGbGRXZn9XCAHjGtwstgj31R1Q).

The intended SWIG way to use the standard library containers is to instantiate them using the `%template` directive. Unfortunately, this runs in a few problems in our build environment:

- `swig_post_process.py` script changes visibility of any entity containing `Internal` in its name to `internal` (see https://cs.corp.google.com/piper///depot/google3/firebase/app/client/unity/swig_post_process.py?rcl=267221433&l=438). This breaks the C# code that SWIG generates for `std::vector`;

- the older version of SWIG we're using doesn't support `std::unordered_map`.

To work around this, C++ wrapper functions and classes are used to ensure that `std::vector` and `std::unordered_map` objects are never actually created in the C# layer. For example, the C# code would create a `FieldValueMap` (which is a simple non-template class) which will get converted into the underlying `std::unordered_map` in the C++ layer.

Running "Leaks" instrument in XCode gives the following numbers when running a somewhat-outdated version of the Unity testapp (containing 29 tests):
- ~7.1K leaks baseline;
- ~4.5K leaks when enabling destructors for Firestore public C++ classes;
- 69 leaks with this CL, a 99% reduction.

Fixing the remaining leaks will be done in a follow-up.

PiperOrigin-RevId: 314847898