wait for SignOut to delete persistent data
We noticed race condition (specifically on macs) where persistent auth data
(in Keystore) isn't deleted before the next test picks up in the suite.
If the next test involves a SignIn and if persistent cache already has a valid
user logged in, there is no change in auth or id tokens and our expected listeners
are not triggered. We could possibly run into this on other platforms too, hence
forcefully waiting 200ms after a SignOut for the persistent data to be cleared.