Add tools for C++ integration tests
Adds the tools needed to build and run C++ integration tests.

These derive from the corresponding tools in google3, with some changes and exceptions:

build_testapps.py and build_desktop_testapps.py have been merged into a single, simpler build_testapps.py.

restore_secrets.py is a new script that decrypts the secret files in scripts/gha-encrypted and restores all the files/values that were stripped out when pushing the integration source code to Github. Except for the reverse ids in the Info.plists needed for iOS, which are not present yet.

xcodebuild.py, provisioning.py and xcode_tool.rb are all for iOS, which is not yet supported until signing is figured out.

This also includes a change to .gitignore to ignore temporary Python files, as well as unencrypted secrets.