Add logic to ignore a specific file in Kokoro build logic
The chromium_certificate_verifier uses a cc file as an include file, which breaks the CMake logic with add_library.  By excluding that file from the add_library, but keeping it in the zip file, it will still be available for the build to work, but will no longer try to treat it as a source file, which caused the problem.

Also, made the local_test script use the regular_grpc flag, since it is now needed

PiperOrigin-RevId: 323644686