Use simple CSV format for OS input, not JSON
Instead of specifying the operating system input as a JSON, this will take a normal CSV string. A separate job, prepare_matrix, will run before the tests to convert the CSV to a JSON list.

The conversion is done by using sed to replace a comma with a quoted comma, and then inserting that value into ["$(new_value)"] using command substitution.