Add tools for C++ integration tests (#103)
* Add tools for C++ integration tests

Adds the tools needed to build and run C++ integration tests.

These derive from the corresponding tools in google3, with some changes and exceptions:

build_testapps.py and build_desktop_testapps.py have been merged into a single, simpler build_testapps.py.

restore_secrets.py is a new script that decrypts the secret files in scripts/gha-encrypted and restores all the files/values that were stripped out when pushing the integration source code to Github. Except for the reverse ids in the Info.plists needed for iOS, which are not present yet.

xcodebuild.py, provisioning.py and xcode_tool.rb are all for iOS, which is not yet supported until signing is figured out.

This also includes a change to .gitignore to ignore temporary Python files, as well as unencrypted secrets.

* Restore reverse id and simplify config reader.

The following changes have been made:

- config_reader.py has been simplified to just return a plain data object containing all the needed config upfront. Thus it's only needed to call one function from the module once.

- Added types to args in docstrings.

- restore_secrets now restores the reverse id in Info.plists, using the value from the decrypted GoogleService-Info.plist as the source of truth.

* Replace argparse with absl.flags

We've decided to use absl.flags as the flag-parsing library for Python in the repo.

Also removed reverse id patching from the xcode tool, since that's being handled by the secret restoration script.