Add integration test workflow (#107)
* Add integration test workflow

Most of this workflow is copied verbatim from the workflow to build the SDK. There are two new parts for the integration tests:

- Manual dispatch with input parameters that can be specified when triggering.
- Two new steps ('Prepare for integration tests' and 'Build and run integration tests').

* Use simple CSV format for OS input, not JSON

Instead of specifying the operating system input as a JSON, this will take a normal CSV string. A separate job, prepare_matrix, will run before the tests to convert the CSV to a JSON list.

The conversion is done by using sed to replace a comma with a quoted comma, and then inserting that value into ["$(new_value)"] using command substitution.