Make changes for Android integration tests
This makes the changes needed to build Android integration tests.

Different platforms are set to run as separate jobs, so that platform-specific issues don't affect other platforms, and to speed things up. This is consistent with the jobs to build SDKs.

The integration test job for Android runs the setup script for Android SDKs, mainly to download the Android NDK.

A non-trivial consideration in the workflows is the caching. It makes sense to use the same caches for the vcpkg dependencies as the SDK build jobs, since those dependencies are the same. But this removes the ccache step for intermediate cmake objects since different cmake targets are being built. We could use separate caches for the integration tests, but we're likely already getting close to the repo limit of 5gb (each configuration for the Android SDK build job uses 500mb by itself). Since the integration tests are not running on PR, speeding these jobs up is less of a priority than the workflows running on PR.

This includes a small change to the restore secrets script, to log a sanitized error message when decryption fails. Previously the error message was suppressed to avoid logging the passphrase, but this makes debugging more difficult. We can just remove the passphrase from the string before logging it.

.gitignore is also changed to ignore intermediate gradle objects that appear in the repo when building for Android.