Use vcpkg to manage toolchains
This attempts to use vcpkg in the same way it's used for the build workflows. There is some duplication with code in the build script for the SDKs, ideally that should be moved to some kind of setup tool that both the build workflows and integration test workflows can run.