Fix issue involving phantom jobs in the workflow
When specifying only a subset of jobs to run, the workflow was nonetheless queueing jobs for the other platforms, with malformed parameters.

This was caused by the "include" statements: their intended use was to supply additional parameters for just the matching operating systems, but when that OS isn't present, include has the effect of adding that OS.

To fix this, the include statements are removed, and the additional variables (in this case, just the triplets) are set as env vars by conditional steps.