Strip whitespace off end of decrypted text.
GPG appears to be introducing a linebreak at the end of decrypted files, which is causing syntax errors in the patched scripts.

i.e. the following:

variable = "PATCH_ME";

is turning into:

variable = "patched_value
";

To fix this, rstrip is used to strip off all whitespace at the end of the decrypted values.