cleanup temp vcpkg directories to save disk space
The buildtrees and downloads directory are used by vcpkg for
building the requested libraries. buildtrees contains all build
intermediates and also the source code for debugging. If you want to
debug with source code for these dependencies, you shouldn't delete
the buildtrees directory. In most cases, these directories are not
needed and can take gigabytes of unnecessary space.