Add Firebase Test Lab tool
This tool sends testapps (apks and ipas) to Firebase Test Lab, and then uses their log output to validate them. This is a heavily modified/simplified version of the tool that exists in google3 (firebase/app/client/unity/testapp_builder/firebase_test_lab).

At a high level the tool works as follows:
- Authorize the use of a Google Cloud project to use its storage bucket.
- Search a directory for apks/ipas.
- Send them to FTL for testing, using the gcloud api. FTL will store their log output in the GCS bucket.
- Once they're done, find the testapp logs in GCS using gsutil, a command line tool installed as part of the Cloud SDK.
- Validate the test results in the testapp logs and provide a summary of all the results.

Changes to the tool in google3 include:
- Significantly less code.
- Makes far fewer assumptions about where FTL puts stuff in cloud storage. No longer requires hard-coded information about FTL's devices (which change over time).
- Cleans up the output.
- Prints the output from FTL when sending testapps, instead of storing it in a file on GCS.
- Prints more helpful information if something unexpected occurs, like being unable to find the testapp log on GCS. This includes a recursive dump of everything added to the GCS bucket for this run, in case anything moved around. This will also make it easier to find items of interest, such as video links. "gsutil cp gs://<etc> <local_path>" can be used to copy the relevant items for further debugging.
- Does not require Linux anymore. Should work on all three platforms, though I haven't tried it on Windows.

Using the tool requires installation of the Google Cloud SDK, which is a requirement of using FTL from the command line. There's a convenient Github action which handles this, though we may add some automation to install this later to simplify local setup.