Work around Android integration test build failures
The workaround is to build twice. We capture the output and check for the presence of the failed task associated with the issue in question. If present, we build again. Otherwise, we raise an error that mimicks the error that would have been raised by running the subprocess with check=True, to maintain the same behaviour. This will make it easier to remove this workaround if/when the underlying issue is fixed.