Add Android tests to integration tests workflow
This modifies the integration test workflow to use the test_lab.py script to send testapps to Firebase Test Lab, and validates the results.

This requires a GCP service account. To authorize its use, we need a key file, which is added to the encrypted files in gha-encrypted. Note that this is not mirrored in google3, since policy prohibits this kind of data from being placed in google3. Instead, the unencrypted file is found in valentine, its valentine id mentioned in the test_lab.py file.

To install the Cloud SDK, we use GCP's setup-gcloud action. There's no particularly convenient way to automate the installation of the Cloud SDK so we use this as a (possibly temporary) convenience. It may be replaced by custom logic so that we can simplify local setup.