Run Android integration tests even on a failed prior step
By default, a step will be skipped if any previous step fails. We override this behaviour for the Android test steps by running on success or failure (but not cancellation). This ensures that in the case where we have 1+ failed builds and 1+ successfuly builds, we still run the tests for the successfully built integration tests.

In the case where all fail to build, or the failure comes from a prior step (so the build step is skipped), the test step will still fail fast. The test_lab.py script is changed to look for testapps before authorizing the GCS service account to fail faster and with less noise.