Add Android integration testing to integration test workflow (#139)
* Add Android tests to integration tests workflow

This modifies the integration test workflow to use the test_lab.py script to send testapps to Firebase Test Lab, and validates the results.

This requires a GCP service account. To authorize its use, we need a key file, which is added to the encrypted files in gha-encrypted. Note that this is not mirrored in google3, since policy prohibits this kind of data from being placed in google3. Instead, the unencrypted file is found in valentine, its valentine id mentioned in the test_lab.py file.

To install the Cloud SDK, we use GCP's setup-gcloud action. There's no particularly convenient way to automate the installation of the Cloud SDK so we use this as a (possibly temporary) convenience. It may be replaced by custom logic so that we can simplify local setup.

* Fix flag names in integration test workflow

*_device -> *_model

* Catch timeouts as normal testapp failures

Currently, build_testapps.py will terminate if any subprocess call times out. Now, timeouts during the building or running of an integration test will be caught and processed as failures, to avoid blocking other integration tests. This is preferred behaviour since e.g. analytics failing should not block remote config from building or running.

Details:
Previously, we were catching subprocess.CalledProcessError, but missing subprocess.TimeoutExpired. These are both instances of the base class subprocess.SubprocessError, so we catch that instead.

* Run Android integration tests even on a failed prior step

By default, a step will be skipped if any previous step fails. We override this behaviour for the Android test steps by running on success or failure (but not cancellation). This ensures that in the case where we have 1+ failed builds and 1+ successfuly builds, we still run the tests for the successfully built integration tests.

In the case where all fail to build, or the failure comes from a prior step (so the build step is skipped), the test step will still fail fast. The test_lab.py script is changed to look for testapps before authorizing the GCS service account to fail faster and with less noise.

* Run tests when not cancelled instead of success+failure.

Instead of explicitly checking for success or failure, just check that we're not cancelled. More readable, and will allow it to run under new statuses that Github may add in the future.